+++ src/test/scala/grizzled/file/GrizzledFileSpec.scala
@@ -14,7 +14,7 @@ class GrizzledFileSpec extends BaseSpec {
     withTemporaryDirectory("GrizzledFile") { dir =>
       val path = new File(fileutil.joinPath(dir.getAbsolutePath, "foobar.txt"))
       path should not (exist)
-      path.touch() should be (success)
+      path.touch() shouldBe 'success
       path should exist
     }
   }
@@ -40,43 +40,43 @@ class GrizzledFileSpec extends BaseSpec {
     withTemporaryDirectory("GrizzledFile") { dir =>
       val absDir = dir.getAbsolutePath
       val f = new File(fileutil.joinPath(absDir, "foo", "bar", "baz.txt"))
-      f.touch() should be (failure)
+      f.touch() shouldBe 'failure
     }
   }
 
   "pathExists" should "return Success for an existing file" in {
     withTemporaryDirectory("GrizzedFile") { dir =>
       val path = new File(fileutil.joinPath(dir.getAbsolutePath, "foo.txt"))
-      path.touch() should be (success)
-      path.pathExists should be (success)
+      path.touch() shouldBe 'success
+      path.pathExists shouldBe 'success
     }
   }
 
   it should "return Failure for a nonexistent file" in {
     withTemporaryDirectory("GrizzedFile") { dir =>
       val path = new File(fileutil.joinPath(dir.getAbsolutePath, "foo.txt"))
-      path.pathExists should be (failure)
+      path.pathExists shouldBe 'failure
     }
   }
 
   "isEmpty" should "return true for an empty directory" in {
     withTemporaryDirectory("GrizzledFile") { dir =>
-      dir.isEmpty should be (true)
+      dir.isEmpty shouldBe true
     }
   }
 
   it should "return false for a non-empty directory" in {
     withTemporaryDirectory("GrizzledFile") { dir =>
       val f = new File(fileutil.joinPath(dir.getAbsolutePath, "foo.txt"))
-      f.touch() should be (success)
-      dir.isEmpty should be (false)
+      f.touch() shouldBe 'success
+      dir.isEmpty shouldBe false
     }
   }
 
   it should "fail with an assertion error for a non-directory" in {
     withTemporaryDirectory("GrizzledFile") { dir =>
       val f = new File(fileutil.joinPath(dir.getAbsolutePath, "foo.txt"))
-      f.touch() should be (success)
+      f.touch() shouldBe 'success
 
       an [AssertionError] should be thrownBy {
         f.isEmpty
@@ -93,10 +93,10 @@ class GrizzledFileSpec extends BaseSpec {
 
       parentDir.mkdirs() should be (true)
       parentDir should exist
-      file.touch() should be (success)
+      file.touch() shouldBe 'success
 
       val t = topDir.deleteRecursively()
-      t should be (success)
+      t shouldBe 'success
       t.get shouldEqual 1
       topDir should not (exist)
       new File(absDir) should exist