+++ src/test/scala/grizzled/util/WithResourceSpec.scala
@@ -7,7 +7,7 @@ import scala.util.Try
 class WithResourceSpec extends BaseSpec {
   private class TestCloseable extends java.io.Closeable with AutoCloseable {
     var isClosed = false
-    def close() = isClosed = true
+    def close(): Unit = isClosed = true
   }
 
   "withResource" should "close a java.io.Closeable on success" in {
@@ -41,7 +41,7 @@ class WithResourceSpec extends BaseSpec {
     import grizzled.util.CanReleaseResource.Implicits.CanReleaseCloseable
 
     val t = tryWithResource(new TestCloseable) { _ => 1 }
-    t shouldBe success
+    t shouldBe 'success
     t.get shouldBe 1
   }
 
@@ -50,7 +50,7 @@ class WithResourceSpec extends BaseSpec {
 
     val c = new TestCloseable
     val t = tryWithResource(c) { _ => 1 }
-    t shouldBe success
+    t shouldBe 'success
     c.isClosed shouldBe true
   }
 
@@ -60,7 +60,7 @@ class WithResourceSpec extends BaseSpec {
     val c = new TestCloseable
     val t = tryWithResource(c) { _ => true }
 
-    t shouldBe success
+    t shouldBe 'success
     c.isClosed shouldBe true
   }
 
@@ -72,12 +72,12 @@ class WithResourceSpec extends BaseSpec {
       throw new java.io.IOException("oops")
     }
 
-    f shouldBe failure
+    f shouldBe 'failure
 
     val msg = f.recover {
       case e: Exception => e.getMessage
     }
-    msg shouldBe success
+    msg shouldBe 'success
     msg.get shouldBe "oops"
   }